fix(storage): fix typo
