+++ src/storage/benches/bench_multi_builder.rs
@@ -26,12 +26,12 @@ use risingwave_common::catalog::TableId;
 use risingwave_hummock_sdk::key::{FullKey, UserKey};
 use risingwave_object_store::object::{ObjectStore, ObjectStoreImpl, S3ObjectStore};
 use risingwave_storage::hummock::multi_builder::{CapacitySplitTableBuilder, TableBuilderFactory};
+use risingwave_storage::hummock::test_utils::{DEFAULT_MAX_KEY_COUNT, DEFAULT_MAX_SST_SIZE};
 use risingwave_storage::hummock::value::HummockValue;
 use risingwave_storage::hummock::{
     BatchSstableWriterFactory, CachePolicy, CompressionAlgorithm, FileCache, HummockResult,
     MemoryLimiter, SstableBuilder, SstableBuilderOptions, SstableStore, SstableWriterFactory,
-    SstableWriterOptions, StreamingSstableWriterFactory, Xor16FilterBuilder, DEFAULT_MAX_KEY_COUNT,
-    DEFAULT_MAX_SST_SIZE,
+    SstableWriterOptions, StreamingSstableWriterFactory, Xor16FilterBuilder,
 };
 use risingwave_storage::monitor::ObjectStoreMetrics;
