+++ src/storage/src/hummock/sstable/builder.rs
@@ -25,18 +25,13 @@ use risingwave_pb::hummock::SstableInfo;
 use super::utils::CompressionAlgorithm;
 use super::{
     BlockBuilder, BlockBuilderOptions, BlockMeta, MonotonicDeleteEvent, SstableMeta, SstableWriter,
-    DEFAULT_BLOCK_SIZE, DEFAULT_ENTRY_SIZE, DEFAULT_RESTART_INTERVAL, VERSION,
+    DEFAULT_ENTRY_SIZE, DEFAULT_RESTART_INTERVAL, VERSION,
 };
 use crate::filter_key_extractor::{FilterKeyExtractorImpl, FullKeyFilterKeyExtractor};
 use crate::hummock::sstable::FilterBuilder;
 use crate::hummock::value::HummockValue;
 use crate::hummock::{HummockResult, MemoryLimiter, Xor16FilterBuilder};
 use crate::opts::StorageOpts;
-
-pub const DEFAULT_SSTABLE_SIZE: usize = 4 * 1024 * 1024;
-pub const DEFAULT_BLOOM_FALSE_POSITIVE: f64 = 0.001;
-pub const DEFAULT_MAX_KEY_COUNT: u64 = 2 * 1024 * 1024;
-pub const DEFAULT_MAX_SST_SIZE: u64 = 512 * 1024 * 1024;
 
 #[derive(Clone, Debug)]
 pub struct SstableBuilderOptions {
@@ -66,20 +61,6 @@ impl From<&StorageOpts> for SstableBuilderOptions {
             compression_algorithm: CompressionAlgorithm::None,
             max_key_count: options.compactor_max_sst_key_count,
             max_sst_size: options.compactor_max_sst_size,
-        }
-    }
-}
-
-impl Default for SstableBuilderOptions {
-    fn default() -> Self {
-        Self {
-            capacity: DEFAULT_SSTABLE_SIZE,
-            block_capacity: DEFAULT_BLOCK_SIZE,
-            restart_interval: DEFAULT_RESTART_INTERVAL,
-            bloom_false_positive: DEFAULT_BLOOM_FALSE_POSITIVE,
-            compression_algorithm: CompressionAlgorithm::None,
-            max_key_count: DEFAULT_MAX_KEY_COUNT,
-            max_sst_size: DEFAULT_MAX_SST_SIZE,
         }
     }
 }
@@ -597,7 +578,7 @@ pub(super) mod tests {
     use crate::hummock::iterator::test_utils::mock_sstable_store;
     use crate::hummock::test_utils::{
         default_builder_opt_for_test, gen_test_sstable_impl, mock_sst_writer, test_key_of,
-        test_value_of, TEST_KEYS_COUNT,
+        test_value_of, DEFAULT_MAX_KEY_COUNT, DEFAULT_MAX_SST_SIZE, TEST_KEYS_COUNT,
     };
     use crate::hummock::{CachePolicy, Sstable, Xor16FilterBuilder, Xor8FilterBuilder};
